+++ packages/web/src/javascripts/Components/ChallengeModal/InputValue.tsx
@@ -2,6 +2,6 @@ import { ChallengePrompt } from '@standardnotes/snjs'
 
 export type InputValue = {
   prompt: ChallengePrompt
-  value: string | number | boolean
+  value: string | number | boolean | Record<string, unknown>
   invalid: boolean
 }