+++ packages/web/src/javascripts/Components/Preferences/Panes/Security/Security.tsx
@@ -11,6 +11,8 @@ import ErroredItems from './ErroredItems'
 import PreferencesPane from '@/Components/Preferences/PreferencesComponents/PreferencesPane'
 import BiometricsLock from '@/Components/Preferences/Panes/Security/BiometricsLock'
 import MultitaskingPrivacy from '@/Components/Preferences/Panes/Security/MultitaskingPrivacy'
+import U2FWrapper from './U2F/U2FWrapper'
+import { featureTrunkEnabled, FeatureTrunkName } from '@/FeatureTrunk'
 
 interface SecurityProps extends MfaProps {
   viewControllerManager: ViewControllerManager
@@ -32,6 +34,9 @@ const Security: FunctionComponent<SecurityProps> = (props) => {
         userProvider={props.userProvider}
         application={props.application}
       />
+      {featureTrunkEnabled(FeatureTrunkName.U2F) && (
+        <U2FWrapper userProvider={props.userProvider} application={props.application} />
+      )}
       {isNativeMobileWeb && <MultitaskingPrivacy application={props.application} />}
       <PasscodeLock viewControllerManager={props.viewControllerManager} application={props.application} />
       {isNativeMobileWeb && <BiometricsLock application={props.application} />}