+++ packages/mobile/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.51.8](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.51.7...@standardnotes/mobile@3.51.8) (2023-02-03)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
+## [3.51.7](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.51.6...@standardnotes/mobile@3.51.7) (2023-02-03)
+
+**Note:** Version bump only for package @standardnotes/mobile
+
 ## [3.51.6](https://github.com/standardnotes/app/compare/@standardnotes/mobile@3.51.5...@standardnotes/mobile@3.51.6) (2023-02-02)
 
 **Note:** Version bump only for package @standardnotes/mobile