+++ packages/services/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.57.0](https://github.com/standardnotes/app/compare/@standardnotes/services@1.56.18...@standardnotes/services@1.57.0) (2023-02-03)
+
+### Features
+
+* **dev:** add u2f ui for managing devices and signing in ([#2182](https://github.com/standardnotes/app/issues/2182)) ([9414774](https://github.com/standardnotes/app/commit/9414774e89c5c21b0c2178652ded72033065e8d2))
+
+## [1.56.18](https://github.com/standardnotes/app/compare/@standardnotes/services@1.56.17...@standardnotes/services@1.56.18) (2023-02-03)
+
+### Bug Fixes
+
+* **web:** disable subscription sharing on demo ([de5973e](https://github.com/standardnotes/app/commit/de5973e15f21fcaff0225834f1c09f295ad83368))
+
 ## [1.56.17](https://github.com/standardnotes/app/compare/@standardnotes/services@1.56.16...@standardnotes/services@1.56.17) (2023-02-02)
 
 **Note:** Version bump only for package @standardnotes/services